requirements.txt
pylint
mypy
flake8
platformio
requests
pyyaml
docker
